refactoring para leitura de linhas
