+++ src/main/java/br/com/pacman/Principal.java
@@ -8,15 +8,7 @@ public class Principal {
 
 	public static void main(String[] args) {
 		
-		try {
-			Scanner scanner = new Scanner(new FileReader("src/main/resources/input.txt"));
-			while (scanner.hasNextLine()) {
-				String linha = scanner.nextLine();
-				System.out.println(linha);
-			}
-		} catch(FileNotFoundException e) {
-			
-		}
+		
 		
 	}
